+++ NETCORE/test/Microsoft.ApplicationInsights.AspNetCore.Tests/Extensions/ApplicationInsightsExtensionsTests/AddApplicationInsightsTelemetryTests.cs
@@ -37,13 +37,21 @@ namespace Microsoft.Extensions.DependencyInjection.Test
 
     public class AddApplicationInsightsTelemetryTests : BaseTestClass
     {
-        [Fact]
-        public static void TelemetryModulesResolvableWhenKeyedServiceRegistered()
+        [Theory]
+        [InlineData(true)]
+        [InlineData(false)]
+        public static void TelemetryModulesResolvableWhenKeyedServiceRegistered(bool manuallyRegisterDiagnosticsTelemetryModule)
         {
             // Note: This test verifies a regression doesn't get introduced for:
-            // https://github.com/dotnet/extensions/issues/5222
+            // https://github.com/microsoft/ApplicationInsights-dotnet/issues/2879
 
             var services = new ServiceCollection();
+
+            services.AddSingleton<ITelemetryModule, TestTelemetryModule>();
+            if (manuallyRegisterDiagnosticsTelemetryModule)
+            {
+                services.AddSingleton<ITelemetryModule, DiagnosticsTelemetryModule>();
+            }
 
             services.AddKeyedSingleton(typeof(ITestService), serviceKey: new(), implementationType: typeof(TestService));
             services.AddKeyedSingleton(typeof(ITestService), serviceKey: new(), implementationInstance: new TestService());
@@ -55,7 +63,9 @@ namespace Microsoft.Extensions.DependencyInjection.Test
 
             var telemetryModules = sp.GetServices<ITelemetryModule>();
 
-            Assert.Equal(7, telemetryModules.Count());
+            Assert.Equal(8, telemetryModules.Count());
+
+            Assert.Single(telemetryModules.Where(m => m.GetType() == typeof(DiagnosticsTelemetryModule)));
         }
 
         [Theory]