+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## VNext
 - [Populate required field Message with "n/a" if it is empty](https://github.com/microsoft/ApplicationInsights-dotnet/issues/1066)
+- [Fix ITelemetryModule singleton registration to support presence of keyed services](https://github.com/microsoft/ApplicationInsights-dotnet/pull/2908)
 
 ## Version 2.22.0
 - no changes since beta.